refactor(connection): move SocketReadStream to private header
