+++ src/connection.cc
@@ -16,10 +16,6 @@
 #include "term.h"
 #include "cursor_p.h"
 
-#include "rapidjson/rapidjson.h"
-#include "rapidjson/encodedstream.h"
-#include "rapidjson/document.h"
-
 namespace RethinkDB {
 
 // constants
@@ -229,74 +225,6 @@ Response ConnectionPrivate::wait_for_response(uint64_t token_want, double wait) 
     ReadLock reader(*this);
     return reader.read_loop(token_want, std::move(guard), wait);
 }
-
-class SocketReadStream {
-public:
-    typedef char Ch;    //!< Character type (byte).
-
-    //! Constructor.
-    /*!
-        \param fp File pointer opened for read.
-        \param buffer user-supplied buffer.
-        \param bufferSize size of buffer in bytes. Must >=4 bytes.
-    */
-    SocketReadStream(ReadLock* reader, char* buffer, size_t bufferSize)
-        : reader_(reader),
-          buffer_(buffer),
-          bufferSize_(bufferSize),
-          bufferLast_(0),
-          current_(buffer_),
-          readCount_(0),
-          count_(0),
-          eof_(false)
-    {
-        RAPIDJSON_ASSERT(reader_ != 0);
-        RAPIDJSON_ASSERT(bufferSize >= 4);
-        Read();
-    }
-
-    Ch Peek() const { return *current_; }
-    Ch Take() { Ch c = *current_; Read(); return c; }
-    size_t Tell() const { return count_ + static_cast<size_t>(current_ - buffer_); }
-
-    // Not implemented
-    void Put(Ch) { RAPIDJSON_ASSERT(false); }
-    void Flush() { RAPIDJSON_ASSERT(false); }
-    Ch* PutBegin() { RAPIDJSON_ASSERT(false); return 0; }
-    size_t PutEnd(Ch*) { RAPIDJSON_ASSERT(false); return 0; }
-
-    // For encoding detection only.
-    const Ch* Peek4() const {
-        return (current_ + 4 <= bufferLast_) ? current_ : 0;
-    }
-
-private:
-    void Read() {
-        if (current_ < bufferLast_) {
-            ++current_;
-        } else if (!eof_) {
-            count_ += readCount_;
-            readCount_ = reader_->recv_some(buffer_, bufferSize_, FOREVER);
-            bufferLast_ = buffer_ + readCount_ - 1;
-            current_ = buffer_;
-
-            if ((count_ + readCount_) == bufferSize_) {
-                buffer_[readCount_] = '\0';
-                ++bufferLast_;
-                eof_ = true;
-            }
-        }
-    }
-
-    ReadLock* reader_;
-    Ch *buffer_;
-    size_t bufferSize_;
-    Ch *bufferLast_;
-    Ch *current_;
-    size_t readCount_;
-    size_t count_;  //!< Number of characters read
-    bool eof_;
-};
 
 Response ReadLock::read_loop(uint64_t token_want, CacheLock&& guard, double wait) {
     if (!guard.inner_lock) {
